🔧[Rename] : TicketManager 파일 분리
